+++ json_serializable/test/integration/json_test_example.dart
@@ -19,6 +19,8 @@ class Person {
 
   Order order;
 
+  MyList<Order> customOrders;
+
   Map<String, Category> houseMap;
   Map<Category, int> categoryCounts;