+++ json_serializable/test/integration/json_test_example.non_nullable.wrapped.dart
@@ -34,6 +34,8 @@ class Person {
 
   Order order;
 
+  MyList<Order> customOrders;
+
   Map<String, Category> houseMap;
   Map<Category, int> categoryCounts;