+++ json_serializable/test/integration/json_test_example.wrapped.dart
@@ -27,6 +27,8 @@ class Person {
 
   Order order;
 
+  MyList<Order> customOrders;
+
   Map<String, Category> houseMap;
   Map<Category, int> categoryCounts;